I have a viewModel with a onpropertychange working
