+++ BaseClasses/BaseViewModel.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// Base class for modeling a view that supports change notifications.
     /// </summary>
-    public abstract class ViewModelBaseClass : ObservableObject
+    public abstract class BaseViewModel : ObservableObject
     {
     }
 }
