+++ MauiCodeSnippets.csproj
@@ -70,7 +70,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <MauiXaml Update="Views\BirthdayPage.xaml">
+	  <MauiXaml Update="MvvmExample\BirthdayPage.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
 	  <MauiXaml Update="Views\CoolButtonPage.xaml">
@@ -88,6 +88,10 @@
 	  <MauiXaml Update="Views\SimpleVVMwithTimerPage.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
+	</ItemGroup>
+
+	<ItemGroup>
+	  <Folder Include="Models\" />
 	</ItemGroup>
 
 </Project>
