+++ MvvmExample/BirthdayPage.xaml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="MauiCodeSnippets.MvvmExample.BirthdayPage"
+             xmlns:viewModel="clr-namespace:MauiCodeSnippets.MvvmExample"
+             Title="BirthdayPage">
+
+    <ContentPage.BindingContext>
+        <viewModel:BirthdayViewModel />
+    </ContentPage.BindingContext>
+
+    <VerticalStackLayout Padding="12">
+        <Label>Resources</Label>
+        <Border>
+            <HorizontalStackLayout Spacing="12" HorizontalOptions="Center">
+                <VerticalStackLayout>
+                    <Entry HorizontalTextAlignment="Center" Text="{Binding MyVar,Mode=TwoWay}" />
+                </VerticalStackLayout>
+            </HorizontalStackLayout>
+        </Border>
+    </VerticalStackLayout>
+</ContentPage>
