+++ MvvmExample/BirthdayPage.xaml.cs
@@ -1,0 +1,9 @@
+namespace MauiCodeSnippets.MvvmExample;
+
+public partial class BirthdayPage : ContentPage
+{
+	public BirthdayPage()
+	{
+		InitializeComponent();
+	}
+}
