--- Views/BirthdayPage.xaml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="MauiCodeSnippets.Views.BirthdayPage"
-             xmlns:vm="clr-namespace:MauiCodeSnippets.ViewModels"
-             Title="BirthdayPage"
-             >
-
-    <!--The Movie Database example from Microsoft has the list of movies as a property in the MainAPp ... did not know you could do this 
-        but I don't want that. I am therefore loading it in the code as it wouldn't work here, and I would get a compiler error -->
-    <!--<ContentPage.BindingContext>
-        <vm:BirthdayListViewModel />
-    </ContentPage.BindingContext>-->
-    
-    <!--Stack Layout allows me to add different elements -->
-    <StackLayout>
-        <Grid>
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="*" />
-                <ColumnDefinition Width="Auto" />
-            </Grid.ColumnDefinitions>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-            </Grid.RowDefinitions>
-
-            <Label Text="Jake" Grid.Row="0" Grid.Column="0" />
-            <Label Text="Add" Grid.Row="0" Grid.Column="1" />
-
-            <Label Text="Penny" Grid.Row="1" Grid.Column="1" />
-            <Entry HorizontalTextAlignment="Center" Grid.Row="1" Grid.Column="1" Text="{Binding Name, Mode=TwoWay}"/>
-
-
-            <ListView ItemsSource="{Binding BirthdayList}" Margin="10" Grid.Row="1" Grid.Column="0">
-                <ListView.ItemTemplate>
-                    <DataTemplate>
-                        <TextCell Text="Mark" x:DataType="vm:BirthdayViewModel">
-                            <TextCell.ContextActions>
-                                <MenuItem Text="Edit" Clicked="OnEditClicked" />
-                                <!--FInd out what IsDestructuve means-->
-                                <MenuItem Text="Delete" Clicked="OnDeleteClicked" IsDestructive="True"/>
-                            </TextCell.ContextActions>
-                        </TextCell>
-                    </DataTemplate>
-                </ListView.ItemTemplate>
-            </ListView>
-        </Grid>
-        
-    </StackLayout>
-</ContentPage>
